+++ khoros/auth.py
@@ -22,6 +22,9 @@ logger = log_utils.initialize_logging(__name__)
 def get_session_key(khoros_object, username=None, password=None):
     """This function retrieves the session key for an authentication session.
 
+    .. versionchanged:: 5.0.0
+       The error handling has been improved to provide more useful information when applicable.
+
     .. versionchanged:: 4.2.0
        The URI is now generated utilizing the :py:func:`khoros.auth._get_khoros_login_url` function.
 
@@ -68,10 +71,18 @@ def get_session_key(khoros_object, username=None, password=None):
         raise errors.exceptions.SessionAuthenticationError()
     else:
         response = response.json()
-        try:
-            session_key = response['response']['value']['$']
-        except KeyError:
-            raise errors.exceptions.SessionAuthenticationError(f"Failed to retrieve a session key for '{username}'")
+        if isinstance(response, dict) and 'response' in response and response['response'].get('status') == 'error':
+            try:
+                error_msg = response['response']['error']['message']
+                full_error_msg = f"Session key authentication failed for '{username}': {error_msg}"
+                raise errors.exceptions.SessionAuthenticationError(full_error_msg)
+            except KeyError:
+                raise errors.exceptions.SessionAuthenticationError(f"Failed to retrieve a session key for '{username}'")
+        else:
+            try:
+                session_key = response['response']['value']['$']
+            except KeyError:
+                raise errors.exceptions.SessionAuthenticationError(f"Failed to retrieve a session key for '{username}'")
     return session_key