+++ docs/changelog.rst
@@ -51,6 +51,12 @@ General
 Changed
 =======
 
+Primary Modules
+---------------
+Changes to the :doc:`primary modules <primary-modules>`.
+
+* Improved the error handling in the :py:func:`khoros.auth.get_session_key` function.
+
 Supporting Modules
 ------------------
 Changes to the :doc:`supporting modules <supporting-modules>`.