+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Change Log
 This page documents the additions, changes, fixes, deprecations and removals made in each release.
 
 **********
-v5.0.0dev2
+v5.0.0dev3
 **********
 **Release Date: TBD**